suddenly realized sometimes a player must pass. ¯\_(ツ)_/¯
this impacts the design a bit. i had been laboring under a
misapprehension: while it is indeed never possible for there to be no
legal moves in a given game of Hive (some pieces is always free), it is
perfectly possible for one player to control none of the free pieces,
forcing them to pass.  also, this is the only situation in which passing
is legal.

altered the move data types to have a Pass alternative. not sure how
much i like that.

it's also awkward with the current design of the Game data structure,
which has no way to represent passing as a possible move.  but i guess
we'd just auto skip to the next turn anyway. though we'd do something in
the UI to notify the players, so we still need to detect this situation.